Extract api_controller's methods for rescue to concerns/rescuable.rb
