+++ app/controllers/concerns/loggable.rb
